+++ transports/daily/src/transport.ts
@@ -3,6 +3,7 @@ import Daily, {
   DailyEventObjectAppMessage,
   DailyEventObjectAvailableDevicesUpdated,
   DailyEventObjectLocalAudioLevel,
+  DailyEventObjectNonFatalError,
   DailyEventObjectParticipant,
   DailyEventObjectParticipantLeft,
   DailyEventObjectRemoteParticipantsAudioLevel,
@@ -157,6 +158,18 @@ export class DailyTransport extends Transport {
     return this._daily.localVideo();
   }
 
+  public enableScreenShare(enable: boolean) {
+    if (enable) {
+      this._daily.startScreenShare();
+    } else {
+      this._daily.stopScreenShare();
+    }
+  }
+
+  public get isSharingScreen(): boolean {
+    return this._daily.localScreenAudio() || this._daily.localScreenVideo();
+  }
+
   tracks() {
     const participants = this._daily?.participants() ?? {};
     const bot = participants?.[this._botId];
@@ -164,6 +177,8 @@ export class DailyTransport extends Transport {
     const tracks: Tracks = {
       local: {
         audio: participants?.local?.tracks?.audio?.persistentTrack,
+        screenAudio: participants?.local?.tracks?.screenAudio?.persistentTrack,
+        screenVideo: participants?.local?.tracks?.screenVideo?.persistentTrack,
         video: participants?.local?.tracks?.video?.persistentTrack,
       },
     };
@@ -281,6 +296,7 @@ export class DailyTransport extends Transport {
     );
     this._daily.on("app-message", this.handleAppMessage.bind(this));
     this._daily.on("left-meeting", this.handleLeftMeeting.bind(this));
+    this._daily.on("nonfatal-error", this.handleNonFatalError.bind(this));
   }
 
   async disconnect() {
@@ -338,17 +354,39 @@ export class DailyTransport extends Transport {
   }
 
   private handleTrackStarted(ev: DailyEventObjectTrack) {
-    this._callbacks.onTrackStarted?.(
-      ev.track,
-      ev.participant ? dailyParticipantToParticipant(ev.participant) : undefined
-    );
+    if (ev.type === "screenAudio" || ev.type === "screenVideo") {
+      this._callbacks.onScreenTrackStarted?.(
+        ev.track,
+        ev.participant
+          ? dailyParticipantToParticipant(ev.participant)
+          : undefined
+      );
+    } else {
+      this._callbacks.onTrackStarted?.(
+        ev.track,
+        ev.participant
+          ? dailyParticipantToParticipant(ev.participant)
+          : undefined
+      );
+    }
   }
 
   private handleTrackStopped(ev: DailyEventObjectTrack) {
-    this._callbacks.onTrackStopped?.(
-      ev.track,
-      ev.participant ? dailyParticipantToParticipant(ev.participant) : undefined
-    );
+    if (ev.type === "screenAudio" || ev.type === "screenVideo") {
+      this._callbacks.onScreenTrackStopped?.(
+        ev.track,
+        ev.participant
+          ? dailyParticipantToParticipant(ev.participant)
+          : undefined
+      );
+    } else {
+      this._callbacks.onTrackStopped?.(
+        ev.track,
+        ev.participant
+          ? dailyParticipantToParticipant(ev.participant)
+          : undefined
+      );
+    }
   }
 
   private handleParticipantJoined(ev: DailyEventObjectParticipant) {
@@ -398,6 +436,14 @@ export class DailyTransport extends Transport {
     this.state = "disconnecting";
     this._botId = "";
     this._callbacks.onDisconnected?.();
+  }
+
+  private handleNonFatalError(ev: DailyEventObjectNonFatalError) {
+    switch (ev.type) {
+      case "screen-share-error":
+        this._callbacks.onScreenShareError?.(ev.errorMsg);
+        break;
+    }
   }
 }