+++ transports/daily/CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-All notable changes to **RTVI Daily Transport** will be documented in this file.
+All notable changes to **Pipecat Daily WebRTC Transport** will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.3.4] - 2024-12-16
+
+### Added
+
+- Screen sharing support
+    - Added `startScreenShare` and `stopScreenShare` methods
+    - Added `isSharingScreen` getter property
 
 ## [0.3.3] - 2024-12-11